+++ examples/input.rs
@@ -1,4 +1,6 @@
 use cushy::figures::units::Px;
+use cushy::styles::components::HorizontalAlignment;
+use cushy::styles::HorizontalAlign;
 use cushy::value::Dynamic;
 use cushy::widget::MakeWidget;
 use cushy::widgets::input::{InputValue, MaskedString};
@@ -14,7 +16,10 @@ fn main() -> cushy::Result {
         .and(password.into_input())
         .into_rows()
         .width(Px::new(100)..Px::new(800))
-        .scroll()
+        .with_local(&HorizontalAlignment, HorizontalAlign::Center)
+        .expand_horizontally()
+        .pad()
+        .vertical_scroll()
         .centered()
         .run()
 }