+++ examples/wrap.rs
@@ -1,8 +1,9 @@
 use cushy::figures::units::Lp;
-use cushy::styles::components::{LineHeight, TextSize};
+use cushy::styles::components::{LineHeight, TextSize, VerticalAlignment};
+use cushy::styles::VerticalAlign;
 use cushy::value::Dynamic;
 use cushy::widget::{MakeWidget, WidgetList};
-use cushy::widgets::wrap::{VerticalAlign, WrapAlign};
+use cushy::widgets::wrap::WrapAlign;
 use cushy::Run;
 use rand::{thread_rng, Rng};
 
@@ -57,8 +58,8 @@ fn main() -> cushy::Result {
                 )
                 .and(
                     vertical_align
-                        .new_radio(VerticalAlign::Middle)
-                        .labelled_by("Middle"),
+                        .new_radio(VerticalAlign::Center)
+                        .labelled_by("Center"),
                 )
                 .and(
                     vertical_align
@@ -76,7 +77,7 @@ fn main() -> cushy::Result {
             words
                 .into_wrap()
                 .align(align)
-                .vertical_align(vertical_align)
+                .with(&VerticalAlignment, vertical_align)
                 .expand_horizontally()
                 .contain()
                 .pad()